subset.nb: Also subset IDs in names attribute
When using IDs in an nb object subsetting leads to an error

> names attribute [old length] must be same length as vector [new length]

This is because subset.nb uses temporary IDs throughout and at the end
overwrites these with the IDs chosen by the user (if set). These are
however not subsetted an therefore we try to overwrite the 'names' attr
of [new length] with user IDs of [old length].

Commit adds an if-clause to the attribute setting that subsets the IDs.